+++ src/renderer/entities/governance/ui/TrackInfo.tsx
