+++ src/renderer/entities/governance/ui/ValueIndicator.tsx
