+++ src/renderer/features/governance/model/unlock/confirm-model.ts
@@ -1,148 +1,14 @@
-import { BN } from '@polkadot/util';
-import { combine, createEvent, createStore, restore, sample } from 'effector';
-import { createForm } from 'effector-forms';
+import { combine, createStore, sample } from 'effector';
 import { spread } from 'patronum';
 
-import { type UnlockChunk } from '@shared/api/governance';
-import {
-  type Account,
-  type Asset,
-  type Chain,
-  type MultisigTxWrapper,
-  type ProxyTxWrapper,
-  type Wallet,
-} from '@shared/core';
-import { ZERO_BALANCE } from '@shared/lib/utils';
-import { networkModel, networkUtils } from '@/entities/network';
-import { transactionBuilder, transactionService } from '@/entities/transaction';
+import { type Asset, type Chain, type MultisigTxWrapper, type ProxyTxWrapper, type Wallet } from '@shared/core';
+import { networkModel } from '@/entities/network';
+import { transactionService } from '@/entities/transaction';
 import { accountUtils, walletModel, walletUtils } from '@entities/wallet';
-import { UnlockRules } from '../../lib/unlock-rules';
-
-type Input = {
-  id?: number;
-  chain: Chain;
-  asset: Asset;
-  unlockableClaims: UnlockChunk[];
-  amount: string;
-};
-
-type BalanceMap = { balance: string; withdraw: string };
-
-type FormParams = {
-  shards: Account[];
-  signatory: Account;
-  amount: string;
-  description: string;
-};
-
-const formInitiated = createEvent<Input>();
-const formSubmitted = createEvent();
-
-const feeChanged = createEvent<string>();
-const totalFeeChanged = createEvent<string>();
-const multisigDepositChanged = createEvent<string>();
-const isFeeLoadingChanged = createEvent<boolean>();
 
 const $networkStore = createStore<{ chain: Chain; asset: Asset } | null>(null);
 const $isMultisig = createStore<boolean>(false);
 const $isProxy = createStore<boolean>(false);
-
-const $accountsBalances = createStore<BalanceMap[]>([]);
-const $signatoryBalance = createStore<string>(ZERO_BALANCE);
-const $proxyBalance = createStore<string>(ZERO_BALANCE);
-
-const $fee = restore(feeChanged, ZERO_BALANCE);
-const $multisigDeposit = restore(multisigDepositChanged, ZERO_BALANCE);
-const $isFeeLoading = restore(isFeeLoadingChanged, true);
-
-const $confirmForm = createForm<FormParams>({
-  fields: {
-    shards: {
-      init: [] as Account[],
-      rules: [
-        {
-          name: 'noProxyFee',
-          source: combine({
-            fee: $fee,
-            isProxy: $isProxy,
-            proxyBalance: $proxyBalance,
-          }),
-          validator: (_s, _f, { isProxy, proxyBalance, fee }) => {
-            if (!isProxy) return true;
-
-            return new BN(fee).lte(new BN(proxyBalance));
-          },
-        },
-      ],
-    },
-    signatory: {
-      init: {} as Account,
-      rules: [
-        {
-          name: 'noSignatorySelected',
-          errorText: 'transfer.noSignatoryError',
-          source: $isMultisig,
-          validator: (signatory, _, isMultisig) => {
-            if (!isMultisig) return true;
-
-            return Object.keys(signatory).length > 0;
-          },
-        },
-        {
-          name: 'notEnoughTokens',
-          errorText: 'proxy.addProxy.notEnoughMultisigTokens',
-          source: combine({
-            fee: $fee,
-            isMultisig: $isMultisig,
-            multisigDeposit: $multisigDeposit,
-            signatoryBalance: $signatoryBalance,
-          }),
-          validator: (_s, _f, { fee, isMultisig, signatoryBalance, multisigDeposit }) => {
-            if (!isMultisig) return true;
-
-            return new BN(multisigDeposit).add(new BN(fee)).lte(new BN(signatoryBalance));
-          },
-        },
-      ],
-    },
-    amount: {
-      init: '',
-      rules: [
-        {
-          name: 'required',
-          errorText: 'transfer.requiredAmountError',
-          validator: Boolean,
-        },
-        {
-          name: 'notZero',
-          errorText: 'transfer.notZeroAmountError',
-          validator: (value) => value !== ZERO_BALANCE,
-        },
-        {
-          name: 'insufficientBalanceForFee',
-          errorText: 'transfer.notEnoughBalanceForFeeError',
-          source: combine({
-            fee: $fee,
-            isMultisig: $isMultisig,
-            accountsBalances: $accountsBalances,
-          }),
-          validator: (value, form, { fee, isMultisig, accountsBalances }) => {
-            if (isMultisig) return true;
-
-            return form.shards.every((_: Account, index: number) => {
-              return new BN(fee).lte(new BN(accountsBalances[index].balance));
-            });
-          },
-        },
-      ],
-    },
-    description: {
-      init: '',
-      rules: [UnlockRules.description.maxLength],
-    },
-  },
-  validateOn: ['submit'],
-});
 
 // Computed
 const $shards = combine(
@@ -271,18 +137,6 @@ const $proxiedWallet = combine(
 //   },
 // );
 
-const $isChainConnected = combine(
-  {
-    network: $networkStore,
-    statuses: networkModel.$connectionStatuses,
-  },
-  ({ network, statuses }) => {
-    if (!network) return false;
-
-    return networkUtils.isConnectedStatus(statuses[network.chain.chainId]);
-  },
-);
-
 const $api = combine(
   {
     apis: networkModel.$apis,
@@ -293,65 +147,6 @@ const $api = combine(
   },
   { skipVoid: false },
 );
-
-const $pureTxs = combine(
-  {
-    network: $networkStore,
-    form: $confirmForm.$values,
-    isConnected: $isChainConnected,
-  },
-  ({ network, form, isConnected }) => {
-    if (!network || !isConnected) return undefined;
-
-    return form.shards.map((shard) => {
-      return transactionBuilder.buildWithdraw({
-        chain: network.chain,
-        accountId: shard.accountId,
-      });
-    });
-  },
-  { skipVoid: false },
-);
-
-const $transactions = combine(
-  {
-    apis: networkModel.$apis,
-    networkStore: $networkStore,
-    pureTxs: $pureTxs,
-    txWrappers: $txWrappers,
-  },
-  ({ apis, networkStore, pureTxs, txWrappers }) => {
-    if (!networkStore || !pureTxs) return undefined;
-
-    return pureTxs.map((tx) =>
-      transactionService.getWrappedTransaction({
-        api: apis[networkStore.chain.chainId],
-        addressPrefix: networkStore.chain.addressPrefix,
-        transaction: tx,
-        txWrappers,
-      }),
-    );
-  },
-  { skipVoid: false },
-);
-
-sample({
-  clock: formInitiated,
-  target: $confirmForm.reset,
-});
-
-sample({
-  clock: formInitiated,
-  source: $shards,
-  filter: (shards) => shards.length > 0,
-  target: $confirmForm.fields.shards.onChange,
-});
-
-sample({
-  clock: formInitiated,
-  fn: ({ chain, asset }) => ({ chain, asset }),
-  target: $networkStore,
-});
 
 sample({
   clock: $txWrappers.updates,
@@ -365,37 +160,14 @@ sample({
   }),
 });
 
-const $canSubmit = combine(
-  {
-    isFormValid: $confirmForm.$isValid,
-    isFeeLoading: $isFeeLoading,
-  },
-  ({ isFormValid, isFeeLoading }) => {
-    return isFormValid && !isFeeLoading;
-  },
-);
-
 export const confirmModel = {
   $api,
   $networkStore,
   $isMultisig,
-  $canSubmit,
+  $isProxy,
   $initiatorWallets,
   $txWrappers,
-  $confirmForm,
   $shards,
   $proxiedWallet,
   $realAccounts,
-  $transactions,
-
-  events: {
-    formInitiated,
-    feeChanged,
-    totalFeeChanged,
-    multisigDepositChanged,
-    isFeeLoadingChanged,
-  },
-  output: {
-    formSubmitted,
-  },
 };
