+++ src/renderer/shared/ui/index.ts
@@ -50,7 +50,6 @@ export { Truncate } from './Truncate/Truncate';
 export { Countdown } from './Countdown/Countdown';
 export { Tabs } from './Tabs/Tabs';
 export { Markdown } from './Markdown/Markdown';
-export { ValueIndicator } from './ValueIndicator/ValueIndicator';
 
 // FIXME: Animation component exported separately.
 // Adding them to this file causes to crash all tests which use anything from that file
